Detach audits and add AuditsMixins
