+++ gvm/protocols/gmpv208/__init__.py
@@ -37,6 +37,7 @@ from gvm.protocols.gmpv208.entities.alerts import (
     get_alert_event_from_string,
     get_alert_method_from_string,
 )
+from gvm.protocols.gmpv208.entities.audits import AuditsMixin
 from gvm.protocols.gmpv208.entities.hosts import (
     HostsMixin,
 )
@@ -84,6 +85,7 @@ PROTOCOL_VERSION = (20, 8)
 class Gmp(
     GmpV208Mixin,
     AlertsMixin,
+    AuditsMixin,
     HostsMixin,
     NotesMixin,
     OperatingSystemsMixin,
